Changing to uint16
With SCPI_MAX_COMMANDS > 255,
these silently overflow, causing commands to be dropped.

Changes:
- Member declarations: max_commands, max_tokens, buffer_length,
  codes_size_, tokens_size_, max_special_commands,
  special_codes_size_ → uint16_t
- operator[] signature: byte/uint8_t → uint16_t
- All for-loop iterators (i, j, k) → uint16_t across all
  implementation files

+++ src/Vrekrer_scpi_arrays_code.h
@@ -5,7 +5,7 @@
 // ## SCPI_String_Array member functions ##
 
 ///Add indexing capability.
-char* SCPI_String_Array::operator[](const uint8_t index) const {
+char* SCPI_String_Array::operator[](const uint16_t index) const {
   if (index >= size_) return NULL; //Invalid index
   return values_[index];
 }
